Move convert_materials tests into separate file
